Move quarantine to kasan_quar.c file
